Refactor integration of performance logs
Remove any dependance on moodle functions in the PHP SDK.

The PHP SDK has been updated to make use of a simple logging object
that is supplied by the callee of the SDK. This allows the SDK to be responsible
for gathering the raw performance data. While the callee is responsible for
writing the performance data to a suitable log.